backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.5
flask-migrate==4.0.4
flask-cors==4.0.0
flask-jwt-extended==4.5.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
werkzeug==2.3.6
flask-bcrypt==1.0.1
pytest==7.3.1